+++ src/components/Editor/Hooks/useProjectList.js
@@ -6,12 +6,12 @@ import { loadProjectList } from '../EditorSlice'
 export const useProjectList = (user) => {
   const dispatch = useDispatch();
   const projectListLoaded = useSelector((state) => state.editor.projectListLoaded);
+  const projectIndexCurrentPage = useSelector((state) => state.editor.projectIndexCurrentPage)
 
   useEffect(() => {
-    if(!user || projectListLoaded === 'success') {
-      return;
+    if (user && projectListLoaded === 'idle') {
+      dispatch(loadProjectList({page: projectIndexCurrentPage, accessToken: user.access_token}));
     }
-    dispatch(loadProjectList(user.access_token));
   }, [user, projectListLoaded]);
 };