+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added release notes to `README.md` (#354)
+- Pagination on 'Your projects' page (#338)
 
 ### Fixed